move getUserIdForVaulting from ViewConfig to Controller-Traits
