+++ src/Controller/PayPalVaultingCardController.php
@@ -3,21 +3,21 @@
 namespace OxidSolutionCatalysts\PayPal\Controller;
 
 use OxidEsales\Eshop\Application\Controller\AccountController;
-use OxidEsales\Eshop\Core\Registry;
-use OxidSolutionCatalysts\PayPal\Core\ServiceFactory;
 use OxidSolutionCatalysts\PayPal\Service\ModuleSettings;
+use OxidSolutionCatalysts\PayPal\Traits\AccountControllerTrait;
 use OxidSolutionCatalysts\PayPal\Traits\ServiceContainer;
 
 /**
- * user account menu for saving paypal for purchase later (vaulting without purchase)
+ * user account menu for saving paypal (acdc) for purchase later (vaulting without purchase)
  */
 class PayPalVaultingCardController extends AccountController
 {
     use ServiceContainer;
+    use AccountControllerTrait;
 
     public function render()
     {
-        $this->_aViewData['vaultingUserId'] = $this->getViewConfig()->getUserIdForVaulting();
+        $this->_aViewData['vaultingUserId'] = $this->getUserIdForVaulting();
         $moduleSettings = $this->getServiceFromContainer(ModuleSettings::class);
 
         if ($moduleSettings->isVaultingAllowedForACDC()) {
@@ -25,19 +25,5 @@ class PayPalVaultingCardController extends AccountController
         }
 
         return parent::render();
-    }
-
-    public function deleteVaultedPayment()
-    {
-        $paymentTokenId = Registry::getRequest()->getRequestEscapedParameter("paymentTokenId");
-        $vaultingService = Registry::get(ServiceFactory::class)->getVaultingService();
-
-        if (!$vaultingService->deleteVaultedPayment($paymentTokenId)) {
-            Registry::getUtilsView()->addErrorToDisplay(
-                Registry::getLang()->translateString('OSC_PAYPAL_DELETE_FAILED'),
-                false,
-                true
-            );
-        }
     }
 }
