+++ src/Core/ViewConfig.php
@@ -264,24 +264,6 @@ class ViewConfig extends ViewConfig_parent
         return $showButton;
     }
 
-    public function getUserIdForVaulting(): string
-    {
-        if (!$this->getUser()) {
-            return "";
-        }
-
-        $payPalCustomerId = $this->getUser()->getFieldData("oscpaypalcustomerid");
-
-        if (!$payPalCustomerId) {
-            return "";
-        }
-
-        $vaultingService = Registry::get(ServiceFactory::class)->getVaultingService();
-        $response = $vaultingService->generateUserIdToken($payPalCustomerId);
-
-        return $response["id_token"] ?? "";
-    }
-
     /**
      * get Session Vault Success
      *
